Retrieve missing quest data
Added to the script a check if there's a quests.json file and see if any of the quest data is missing from it, if so only get set quest data instead of it scraping all of them all over again every time a new quest releases.

+++ start.py
@@ -2,6 +2,7 @@ import requests
 import json
 import time
 import re
+import os
 from bs4 import BeautifulSoup
 
 skills = [
@@ -10,32 +11,7 @@ skills = [
     "Firemaking", "Woodcutting", "Slayer"
 ]
 
-# Send a GET request to the web page
-url = "https://oldschool.runescape.wiki/w/Quests/List"
-response = requests.get(url)
-
-# Create a BeautifulSoup object to parse the HTML content
-soup = BeautifulSoup(response.text, "html.parser")
-
-# Find the table containing the quest list
-table = soup.find_all("table", class_="sortable")
-
-# Extract information from the table
-rows = []
-rows.extend(table[0].find_all("tr"))
-rows.extend(table[1].find_all("tr"))
-
-# Get the quest names from the rows
-quests = []
-for row in rows[1:]:
-    cells = row.find_all("td")
-    if len(cells) >= 2:
-        quest_name = cells[1].find("a").text.strip()
-        quests.append(quest_name)
-
-quest_json = []
-
-for i, quest_name in enumerate(quests, start=1):
+def wiki_scraper(quest_name, data):
     url = f"https://oldschool.runescape.wiki/w/{quest_name.replace(' ', '_')}"
     response = requests.get(url)
     soup = BeautifulSoup(response.text, "html.parser")
@@ -115,16 +91,72 @@ for i, quest_name in enumerate(quests, start=1):
         "requirements": requirements,
         "rewards": rewards
     }
-    quest_json.append(quest_data)
+    data.append(quest_data)
 
-    percentage = (i / len(quests)) * 100
-    print(f"{i}/{len(quests)} ({percentage:.2f}%)")
+    return data
 
-    # This is just here to not overload the OSRS WIKI with too much traffic at once, idk if this makes any difference though... xd
-    if i < len(quests):
-        time.sleep(3)
+# Send a GET request to the web page
+url = "https://oldschool.runescape.wiki/w/Quests/List"
+response = requests.get(url)
 
-with open("quests.json", "w") as file:
-    json.dump(quest_json, file)
+# Create a BeautifulSoup object to parse the HTML content
+soup = BeautifulSoup(response.text, "html.parser")
 
-print("DONE")
+# Find the table containing the quest list
+table = soup.find_all("table", class_="sortable")
+
+# Extract information from the table
+rows = []
+rows.extend(table[0].find_all("tr"))
+rows.extend(table[1].find_all("tr"))
+
+# Get the quest names from the rows
+quests = []
+for row in rows[1:]:
+    cells = row.find_all("td")
+    if len(cells) >= 2:
+        quest_name = cells[1].find("a").text.strip()
+        quests.append(quest_name)
+
+quest_json = []
+data = []
+
+file = "quests.json"
+
+if os.path.exists(file):
+    with open(file, 'r') as file:
+        data = json.load(file)
+
+    data_names = []
+    for i in range(len(data)):
+        data_names.append(data[i].get("name"))
+
+    set1 = set(quests)
+    set2 = set(data_names)
+
+    # Find the elements missing in array2 (present in array1 but not in array2)
+    missing_entries = set1.difference(set2)
+
+    if missing_entries:
+        for quest in missing_entries:
+            print("Retrieving missing quest data for - '" + quest + "'")
+            quest_json = wiki_scraper(quest, data)
+            time.sleep(3)
+
+        with open("quests.json", "w") as file:
+            json.dump(quest_json, file)
+    else:
+        print("No quest data is missing...")
+else:
+    for i, quest_name in enumerate(quests, start=1):
+        quest_json.append(wiki_scraper(quest_name, data=[]))
+
+        percentage = (i / len(quests)) * 100
+        print(f"{i}/{len(quests)} ({percentage:.2f}%)")
+
+        # This is just here to not overload the OSRS WIKI with too much traffic at once, idk if this makes any difference though... xd
+        if i < len(quests):
+            time.sleep(3)
+
+    with open("quests.json", "w") as file:
+        json.dump(quest_json, file)
